Web: Enable PowerSync extensions even without worker
