+++ packages/powersync_core/lib/src/open_factory/web/web_open_factory.dart
@@ -3,14 +3,28 @@ import 'dart:async';
 import 'package:powersync_core/src/open_factory/abstract_powersync_open_factory.dart';
 import 'package:powersync_core/src/uuid.dart';
 import 'package:sqlite_async/sqlite3_common.dart';
+import 'package:sqlite_async/sqlite3_web.dart';
 import 'package:sqlite_async/sqlite_async.dart';
+import 'package:sqlite_async/web.dart';
+
+import '../../web/worker_utils.dart';
 
 /// Web implementation for [AbstractPowerSyncOpenFactory]
-class PowerSyncOpenFactory extends AbstractPowerSyncOpenFactory {
+class PowerSyncOpenFactory extends AbstractPowerSyncOpenFactory
+    implements WebSqliteOpenFactory {
   PowerSyncOpenFactory({
     required super.path,
     super.sqliteOptions,
   });
+
+  @override
+  Future<WebSqlite> openWebSqlite(WebSqliteOptions options) async {
+    return WebSqlite.open(
+      wasmModule: Uri.parse(sqliteOptions.webSqliteOptions.wasmUri),
+      worker: Uri.parse(sqliteOptions.webSqliteOptions.workerUri),
+      controller: PowerSyncAsyncSqliteController(),
+    );
+  }
 
   @override
   void enableExtension() {
