+++ packages/powersync_core/lib/src/web/powersync_db.worker.dart
@@ -4,30 +4,10 @@
 
 library;
 
-import 'dart:js_interop';
-
-import 'package:sqlite_async/sqlite3_web_worker.dart';
 import 'package:sqlite_async/sqlite3_web.dart';
-import 'package:sqlite_async/sqlite3_wasm.dart';
 
 import 'worker_utils.dart';
 
 void main() {
   WebSqlite.workerEntrypoint(controller: PowerSyncAsyncSqliteController());
-}
-
-final class PowerSyncAsyncSqliteController extends AsyncSqliteController {
-  @override
-  Future<WorkerDatabase> openDatabase(
-      WasmSqlite3 sqlite3, String path, String vfs) async {
-    final asyncDb = await super.openDatabase(sqlite3, path, vfs);
-    setupPowerSyncDatabase(asyncDb.database);
-    return asyncDb;
-  }
-
-  @override
-  Future<JSAny?> handleCustomRequest(
-      ClientConnection connection, JSAny? request) {
-    throw UnimplementedError();
-  }
 }
